implemented jwt token

+++ backend/controllers/goalController.js
@@ -2,12 +2,13 @@ const { request } = require('express')
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler (async (req,res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
 
     res.status(200).json(goals)
 })
@@ -23,7 +24,8 @@ const setGoal = asyncHandler(async (req,res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal)
@@ -41,6 +43,21 @@ const putGoals = asyncHandler(async (req,res) => {
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    //Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user mathes the goal user
+    if(goal.user.toString() !== user.id)
+    {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
     res.status(200).json(updatedGoal)
@@ -55,6 +72,20 @@ const deleteGoals = asyncHandler(async (req,res) => {
     {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+
+    //Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user mathes the goal user
+    if(goal.user.toString() !== user.id)
+    {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await goal.deleteOne()
